Refactor in identity package
